chore: refactor code (#9)
