+++ auth-service/src/main/java/com/finemeet/authservice/broker/producer/NotificationProducerService.java
@@ -1,6 +1,6 @@
-package com.finemeet.authservice.kafka.producer;
+package com.finemeet.authservice.broker.producer;
 
-import com.finemeet.authservice.kafka.config.KafkaTopicProperties;
+import com.finemeet.authservice.broker.config.KafkaTopicProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -13,8 +13,8 @@ public class NotificationProducerService {
     private final KafkaTemplate<String, String> kafkaTemplate;
     private final KafkaTopicProperties topicProperties;
 
-    public void sendUserRegisteredEvent(com.finemeet.common.notification.NotificationEvent notificationEvent) {
-        String topic = topicProperties.getProducer().getUserRegistered();
+    public void publishNotificationsEvent(com.finemeet.common.notification.NotificationEvent notificationEvent) {
+        String topic = topicProperties.getProducer().getNotifications();
         try {
             kafkaTemplate.send(topic, notificationEvent.getRecipientId(), notificationEvent.toString());
             log.info("Sent message to [{}] -> data={}", topic, notificationEvent);
