+++ notification-service/src/main/java/com/finemeet/notificationservice/broker/NotificationConsumerService.java
@@ -1,0 +1,5 @@
+package com.finemeet.notificationservice.broker;
+
+public class NotificationConsumerService {
+
+}
